♻️ FR13K9 insights: decompose report command
