+++ .agentplane/tasks/202605282300-FR13K9/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 FR13K9 task: Insights command report decomposition [202605282300-FR13K9]
