+++ .agentplane/tasks/202605282300-FR13K9/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605282300-FR13K9/insights-report-decomposition",
+  "created_at": "2026-05-28T23:00:26.850Z",
+  "diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "last_verified_at": "2026-05-28T23:05:30.755Z",
+  "last_verified_diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "schema_version": 1,
+  "task_id": "202605282300-FR13K9",
+  "updated_at": "2026-05-28T23:00:26.850Z",
+  "verify": {
+    "status": "pass"
+  }
+}
