+++ .agentplane/tasks/202605282300-FR13K9/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-28T23:00:26.850Z
+
+## Task
+
+- Task: `202605282300-FR13K9`
+- Title: Insights command report decomposition
+- Status: DOING
+- Branch: `task/202605282300-FR13K9/insights-report-decomposition`
+- Canonical task record: `.agentplane/tasks/202605282300-FR13K9/README.md`
+
+## Verification
+
+- State: ok
+- Note: Insights command report decomposition completed. Verified with insights report/issue/error-map tests, typecheck, arch deps, lint, format, and hotspot threshold check (runtime warnings 37 -> 36).
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T23:00:26.850Z
+- Branch: task/202605282300-FR13K9/insights-report-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
